refactor(graphql): rename account_sep41 → account_allowances and relocate non-allowance tests
