+++ internal/serve/graphql/resolvers/account.resolvers.go
@@ -175,7 +175,7 @@ func (r *accountResolver) StateChanges(ctx context.Context, obj *types.Account, 
 
 // Sep41Allowances is the resolver for the sep41Allowances field. The underlying
 // SQL page is keyset-paginated and the expiration filter is applied server-side;
-// getSEP41Allowances does the heavy lifting (see account_sep41.go).
+// getSEP41Allowances does the heavy lifting (see account_allowances.go).
 func (r *accountResolver) Sep41Allowances(ctx context.Context, obj *types.Account, first *int32, after *string, last *int32, before *string) (*graphql1.SEP41AllowanceConnection, error) {
 	return r.getSEP41Allowances(ctx, string(obj.StellarAddress), first, after, last, before)
 }
